Updated all functions that require a size to take unsigned long instead of int

+++ DataStructures/Common/Node.c
@@ -18,7 +18,7 @@
 // MARK: CONSTRUCTORS
 
 // The constructor is used to create new instances of nodes.
-struct Node node_constructor(void *data, int size)
+struct Node node_constructor(void *data, unsigned long size)
 {
     if (size < 1)
     {
